requirements.txt
Django==1.10
Fabric==1.8.0
django-extensions==1.7.2
lxml==3.5.0
jcconv==0.2.3
gunicorn==19.4.3
redis==2.10.5
hiredis==0.2.0
pytz==2013d
django-allauth==0.18.0
django-nose==1.4.4
django-lazysignup==1.0.2  # TODO Remove?
django-model-utils==2.4
rq==0.5.6
django-rq==0.9.0
rq-dashboard
natto-py==0.7.0
twython==3.3.0
djangorestframework==3.4.6
django-filter==0.14.0
markdown==2.6.5
django-crispy-forms==1.5.2
requests[security]==2.7.0
humanize==0.5.1
python-dateutil==2.5.2
djoser==0.5.1
django-silk==0.7.0

# django-devserver + optional dependencies
https://github.com/dcramer/django-devserver/archive/master.zip
#django_devserver
sqlparse
werkzeug
guppy
line_profiler

bpython
psycopg2
unidecode

https://github.com/kennethreitz/dj-static/archive/e73fa3bce00ac30b78218d60ece9efe931058a95.zip
https://github.com/aehlke/django-cachecow/archive/master.zip
https://github.com/aehlke/django-kanjivg/archive/master.zip
https://github.com/aehlke/django-snippetscream/archive/master.zip

# For django-catnap.
WebOb==1.0.3

# TODO: These should be used, but require some code work:
# django-lazysignup


# Deprecated, but used for migration to v2:
django-timezones==0.2
BeautifulSoup==3.2.0
https://github.com/aehlke/django-catnap/archive/master.zip


# Not sure if used anymore?
rfc3339==5
